requirements.txt
fastapi
uvicorn
supabase
pandas
openai
pypdf
python-dotenv
tiktoken
openpyxl
